Refactor: involve specification into domain layer
Specification definitions in domain layer is as interfaces, and
correspondingly they are implemented in mybatis implementation package
for this project and be used in production code through spring.

Change 2 methods, `fetchResources` and `insertSubresource`, in
ResourceRepository to support specification query and modify
corresponding usage occurred in `ResourceListHandler` and
`MultiUploadCompleteContentHandler`.

Fix `continuation` support in `fetchResources`(`listResources` previously).

Signed-off-by: RichardLea <chigix@zoho.com>