Fix: datasource-upload init errors for executing init sql
Because spring tag `jdbc:initialize-database` would execute the script
unconditionally, while h2database's data dir is checked and initialized
after chizuru application context build, sql errors will surely emitted
if a project has initialized and ran before. This bug is fixed in this
commit by replacing `jdbc:initialize-database` tag with a database
populator bean, whose class is located in spring package, and manually
instantiated in ApplicationContext creation later.

For the modification in ApplicationContext instantiation, another
ApplicationContextFactoryBean in test package for testing is written and
used to replace the reference to original context factory bean class in
source package, which is not designed for testing environment.

Signed-off-by: RichardLea <chigix@zoho.com>